Add the "OnHold" options.
This is really useful for switching between "work" and "home".

If I'm at work and have a browser with 10 tabs related to work and then
go home, I can hit the hotkey and the browser and terimnals are all
moved out of the way and I can come back to it later without giving up
my standard workspaces.